@@ -44,7 +44,7 @@
         "doc:api:datetime": "copyfiles -u 2 datetime/temp/rest-datetime.api.md docs/content/api/ng-rest-datetime && rimraf datetime/temp"
     },
     "engines": {
-        "node": "^14.15.0 || >=16.10.0"
+        "node": "^14.20.0 || ^16.13.0 || >=18.10.0"
     },
     "repository": {
         "type": "git",
@@ -62,39 +62,39 @@
     },
     "homepage": "https://github.com/ressurectit/ng-rest",
     "peerDependencies": {
-        "@angular/core": ">=14.0.0",
-        "@angular/common": ">=14.0.0",
-        "rxjs": ">=6.6.7",
-        "@jscrpt/common": ">=2.2.0",
-        "@anglr/common": ">=11.2.0",
-        "@anglr/datetime": ">=3.0.0",
+        "@angular/core": ">=15.0.0",
+        "@angular/common": ">=15.0.0",
+        "rxjs": ">=7.5.7",
+        "@jscrpt/common": ">=3.3.0",
+        "@anglr/common": ">=15.0.1",
+        "@anglr/datetime": ">=5.0.0",
         "crypto-js": "^4.0.0",
         "@stomp/stompjs": "^5.4.2",
         "sockjs-client": "^1.3.0",
         "avsc": "^5.4.21",
-        "tslib": "^2.3.1"
+        "tslib": "^2.4.1"
     },
     "devDependencies": {
-        "@angular/core": "^14.0.0",
-        "@angular/common": "^14.0.0",
-        "@angular/platform-browser": "^14.0.0",
-        "@angular/animations": "^14.0.0",
-        "@angular/forms": "^14.0.0",
-        "rxjs": "^6.6.7",
-        "@jscrpt/common": "^2.2.0",
-        "@anglr/common": "^11.2.0",
-        "@anglr/datetime": "^3.0.0",
+        "@angular/core": "^15.0.0",
+        "@angular/common": "^15.0.0",
+        "@angular/platform-browser": "^15.0.0",
+        "@angular/animations": "^15.0.0",
+        "@angular/forms": "^15.0.0",
+        "rxjs": "^7.5.7",
+        "@jscrpt/common": "^3.3.0",
+        "@anglr/common": "^15.0.1",
+        "@anglr/datetime": "^5.0.0",
         "avsc": "^5.4.21",
         "@types/node": "^14.14.31",
         "@stomp/stompjs": "^5.4.2",
         "@types/crypto-js": "4.0.2",
         "@types/sockjs-client": "^1.1.1",
-        "@angular/compiler-cli": "^14.0.0",
-        "@angular/compiler": "^14.0.0",
-        "@angular/language-service": "^14.0.0",
+        "@angular/compiler-cli": "^15.0.0",
+        "@angular/compiler": "^15.0.0",
+        "@angular/language-service": "^15.0.0",
         "@microsoft/api-extractor": "7.19.2",
         "@microsoft/api-documenter": "7.13.77",
-        "tslib": "2.3.1",
+        "tslib": "2.4.1",
         "eslint": "7.32.0",
         "@typescript-eslint/eslint-plugin": "4.33.0",
         "@typescript-eslint/parser": "4.33.0",
@@ -103,7 +103,7 @@
         "copyfiles": "2.4.1",
         "replace-in-file": "6.3.2",
         "path-exists-cli": "2.0.0",
-        "typescript": "4.7.2",
+        "typescript": "4.8.4",
         "npm-git-version": "2.0.0"
     },
     "esm2020": "./es2020/src/index.js",
